Fixing endian testing/conversion routines to use unit32_t.
git-svn-id: file:///home/chulwoo/CPS/repo/CVS2SVN/cps/trunk@2985 fb734a29-28db-4bc5-88a3-3cb5ee595545

+++ cps_pp/include/util/intconv.h
@@ -7,8 +7,8 @@
 CPS_START_NAMESPACE
 using namespace std;
 
-typedef unsigned long int type32;
-typedef unsigned long long type64; // maybe later we'll have 64-bit int...
+typedef uint32_t type32;
+typedef uint64_t type64; // maybe later we'll have 64-bit int...
 
 
 enum INT_FORMAT {
